fix(imports): consolidate imports from chaturbate_poller in example scripts
