+++ src/chaturbate_poller/__init__.py
@@ -1,9 +1,10 @@
 """chaturbate_poller package."""
 
-# Read version from installed package
 from importlib.metadata import version
 
 from chaturbate_poller.chaturbate_client import ChaturbateClient
+from chaturbate_poller.config_manager import ConfigManager
+from chaturbate_poller.format_messages import format_message
 
 __version__ = version("chaturbate_poller")
 __author__ = "MountainGod2"
@@ -15,5 +16,4 @@ __url__ = "https://github.com/MountainGod2/chaturbate_poller"
 __description__ = "Python library for interacting with the Chaturbate Events API."
 
 
-__all__ = ["ChaturbateClient"]
-"""List[str]: The package exports."""
+__all__: list[str] = ["ChaturbateClient", "ConfigManager", "format_message"]
